# R401-Architecture_logicielle/pokapi-user/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "pokapi-user API Documentation",
    "description": "",
    "version": "1.0.0"
  },
  "host": "localhost:8082/api/v0",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
    "/login": {
      "post": {
        "summary": "Se connecter à partir des identifiants ou d'un token",
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "login": {
                  "example": "any"
                },
                "password": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/register": {
      "post": {
        "summary": "Inscrire un nouvel utilisateur",
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "login": {
                  "example": "any"
                },
                "pseudo": {
                  "example": "any"
                },
                "password": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/my-cards": {
      "get": {
        "summary": "Récupérer toutes les cartes de la collection de l'utilisateur",
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "503": {
            "description": "Service Unavailable"
          }
        }
      }
    },
    "/my-cards/{cardId}": {
      "get": {
        "summary": "Récupérer une carte de la collection de l'utilisateur",
        "description": "",
        "parameters": [
          {
            "name": "cardId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          },
          "501": {
            "description": "Not Implemented"
          }
        }
      }
    },
    "/pseudo/{pseudo}": {
      "get": {
        "summary": "Trouver un utilisateur à partir de son pseudo",
        "description": "",
        "parameters": [
          {
            "name": "pseudo",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/open-booster/{setId}": {
      "get": {
        "summary": "Ouvrir un booster pour ajouter des cartes dans la collection",
        "description": "",
        "parameters": [
          {
            "name": "setId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          },
          "501": {
            "description": "Not Implemented"
          }
        }
      }
    },
    "/info": {
      "get": {
        "summary": "Récupérer son login et pseudo",
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/update": {
      "put": {
        "summary": "Mettre à jour son pseudo et mot de passe",
        "description": "",
        "responses": {
          "401": {
            "description": "Unauthorized"
          },
          "501": {
            "description": "Not Implemented"
          }
        }
      }
    },
    "/delete": {
      "delete": {
        "summary": "Supprimer l'utilisateur",
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/searched": {
      "get": {
        "summary": "Récupérer toutes les cartes marquées comme recherchées d'un utilisateur",
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/searched/add": {
      "post": {
        "summary": "Ajouter une carte dans la liste des cartes recherchées",
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    }
  }
}